move duration type test
